refactor: use real Calendly demo page, fix card CTA alignment, extract story config
- drop the hand-rolled react-calendly Storybook mock; point CTAs at
  calendly.com/calendly-demo so the genuine widget renders
- pin offer-card pricing/CTA footer to the bottom so buttons align across cards
- extract story args/argTypes/builders into landingPageStoryControls.ts
