+++ src/components/LandingPage/landingPage.scss
@@ -80,6 +80,10 @@ body .calendly-overlay .calendly-popup {
   color: var(--fg-subtle);
 }
 
+.Layer__LandingPage-options__footer {
+  margin-block-start: auto;
+}
+
 .Layer__LandingPage__content {
   display: flex;
   flex-direction: column;
